use a supported unit for snow depth

+++ internal/pkg/application/decoder/payload/payload.go
@@ -223,9 +223,9 @@ func Presence(p bool) PayloadDecoratorFunc {
 	})
 }
 
-func SnowHeight(sh int) PayloadDecoratorFunc {
+func SnowHeight(sh float64) PayloadDecoratorFunc {
 	return S("snowHeight", struct {
-		SnowHeight int
+		SnowHeight float64
 	}{
 		sh,
 	})
